requirements.txt
asgiref==3.5.2
certifi==2022.9.24       
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4        
cryptography==38.0.3     
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.8.0
idna==3.4
importlib-metadata==4.12.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
markdown2==2.4.3
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.2.0
pipenv==2022.11.5
platformdirs==2.5.3
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
virtualenv-clone==0.5.7
zipp==3.8.1